+++ index.php
@@ -2,21 +2,7 @@
 
 <!DOCTYPE html>
 <html>
-<head>
-
-	<meta charset="UTF-8">
-	<meta name="description" content="Sweet Elite is a dating sims inspired by the popular Japanese Otome.Flirt with students and uncover Arlington Academy's dark secrets.Choose your own story!">
-	
-	<title>Sweet Elite: Flirt and Uncover the Secrets of Arlington Academy!</title>
-
-	<link href="css/se-stylesheet.css" rel="stylesheet" type="text/css">
-	<link href='https://fonts.googleapis.com/css?family=Roboto:400,100,100italic,300,300italic,400italic,500,500italic,700,700italic,900,900italic' rel='stylesheet' type='text/css'>
-	<link href="https://fonts.googleapis.com/css?family=Great+Vibes" rel="stylesheet" type='text/css'>
-
-	<link rel="icon" type="image/ico" href="images/favicon.png"/>
-	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
-
-</head>
+<?php include_once 'partials/head.php' ?>
 
 <body>
 
@@ -221,9 +207,9 @@
 				<h4 class="center" style="margin-bottom: 10px;">Welcome Back, <?php if(isset($_SESSION['username'])) echo $_SESSION['username']; ?>!</h4> <!-- <a href="logout.php">Logout</a> </p> -->
 
 				<!--FIRST SECTION: Profile overview & Current Chapter-->
-				<div class="desktop-flex-container">
+				<div class="flex-container">
 					<!--LEFT container: PROFILE INFO-->
-					<div class="desktop-flex-panel logged-in-three-containers">
+					<div class="flex-panel logged-in-three-containers">
 						<!--Profile image-->	
 						<div id="profile-overview-top" class="flex-panel">
 							<img id="index-profile-img" style="width:100%; display: block;" src="profile/character_small_portrait.php"/>
@@ -231,16 +217,14 @@
 						</div>
 
 						<!--Buttons container-->	
-						<div id="profile-overview-bottom" class="center"><!--border: 2px dotted #f3def2; border-top: none;">-->
+						<div id="profile-overview-bottom" class="flex-container"><!--border: 2px dotted #f3def2; border-top: none;">-->
 							<a class="button pink_button" style="margin: 10px;" href="dorm.php">Dorm</a>
 							<a class="button pink_button" style="margin: 10px;" href="profile.php">Profile</a>
-							<div class="flex-panel"></div>
-							<div class="flex-panel"></div>
 						</div>
 					</div>
 
 					<!--MIDDLE container: MOST AFFINITY-->
-					<div class="desktop-flex-panel flex-container-vertical logged-in-three-containers">
+					<div class="flex-panel flex-container-vertical logged-in-three-containers">
 						<div class="flex-panel" style="padding: 10px; border: 2px solid #f3def2; border-top-left-radius: 10px; border-top-right-radius: 10px;">
 							<!--title-->
 							<div class="flex-container">
@@ -273,7 +257,7 @@
 					</div>
 				
 					<!--RIGHT container: CURRENT CHAPTER-->
-					<div class="desktop-flex-panel flex-container-vertical logged-in-three-containers">
+					<div class="flex-panel flex-container-vertical logged-in-three-containers">
 						<div class="flex-panel" style="padding: 10px; border: 2px solid #fdeac7; border-top-left-radius: 10px; border-top-right-radius: 10px;">
 							<h4 id="currentChapter" style="margin: 10px; text-align: center; font-weight: 300; font-size: 1.5em;">Current Chapter:</h4>