+++ partials/footers.php
@@ -25,19 +25,33 @@
 <script src = "js/variables.js"></script>
 <script src = "js/library.js"></script>
 
+<!-- SE Adsense -->
+<!-- Publisher ID: ca-pub-5215030892730474 -->
+<!-- Customer ID: 4063247438 -->
+<ins class="adsbygoogle"
+    style="display:block"
+    data-ad-client="ca-pub-5215030892730474"
+    data-ad-slot="2669137948"
+    data-ad-format="auto"></ins>
+
 <footer>
-	<p>Dulcet Games Inc. © 2016 - 2018. All rights reserved.</p><br>
+    <p>Dulcet Games Inc. © 2016 - 2018. All rights reserved.</p><br>
     <p class="footer_link"><a href="index.php">HOME</a> | <a href="terms.php">TERMS</a> | <a href="shop.php">SHOP</a> | <a href="about.php">ABOUT</a> | <a href="contact.php">CONTACT</a> | <a href="parents.php">FOR PARENTS</a></p>
     
     <!-- SE Google Ads-->
     <script async src="//pagead2.googlesyndication.com/pagead/js/adsbygoogle.js"></script>
-    <!-- SE -->
-    <ins class="adsbygoogle"
-        style="display:block"
-        data-ad-client="ca-pub-5215030892730474"
-        data-ad-slot="2669137948"
-        data-ad-format="auto"></ins>
-    <script>
-    (adsbygoogle = window.adsbygoogle || []).push({});
-    </script>
+    
+    <script>(adsbygoogle = window.adsbygoogle || []).push({});</script>
+
+    <!-- <script>
+        // window.onload = function(){(adsbygoogle = window.adsbygoogle || []).push({});}
+        window.onload = function()
+        {
+            (adsbygoogle = window.adsbygoogle || []).push
+            ({
+                google_ad_client: "ca-pub-5215030892730474",//Google Ads
+                enable_page_level_ads: true
+            });
+        }
+    </script> -->
 </footer>